requirements.txt
fastapi
httpx[socks]
openai
pydantic
pydantic_settings
requests
starlette
uvicorn
google-genai
jinja2
python-multipart
cryptography
pymysql
sqlalchemy
aiomysql
aiosqlite
databases
python-dotenv
apscheduler
packaging
